Refactoring and trying to use iyzico payment

+++ DigitalStore.Api/DigitalStore.Schema/User.cs
@@ -26,6 +26,6 @@ namespace DigitalStore.Schema
         public string Gender { get; set; }
         public string Occupation { get; set; }
         public DateTime? DateOfBirth { get; set; }
-        public decimal? PointsWallet { get; set; }
+        public float? PointsWallet { get; set; }
     }
 }
